Move the Result Transformers to their own file
