+++ packages/neo4j-driver-deno/lib/core/result-eager.ts
@@ -19,7 +19,6 @@
 
 import Record, { Dict } from './record.ts'
 import ResultSummary from './result-summary.ts'
-import Result from './result.ts'
 
 /**
  * Represents the fully streamed result
@@ -38,7 +37,7 @@ export default class EagerResult<Entries extends Dict = Dict> {
    */
   constructor (
     keys: string[],
-    records: Record[],
+    records: Array<Record<Entries>>,
     summary: ResultSummary
   ) {
     /**
@@ -57,18 +56,4 @@ export default class EagerResult<Entries extends Dict = Dict> {
      */
     this.summary = summary
   }
-}
-
-/**
- * Creates an {@link EagerResult} from a given {@link Result} by
- * consuming the whole stream.
- *
- * @private
- * @param {Result} result The result to be consumed
- * @returns A promise of an EagerResult
- */
-export async function createEagerResultFromResult<Entries extends Dict = Dict> (result: Result): Promise<EagerResult<Entries>> {
-  const { summary, records } = await result
-  const keys = await result.keys()
-  return new EagerResult(keys, records, summary)
 }
